+++ main.py
@@ -1,15 +1,13 @@
 from gates import *
 from register import register
-from bus import bus
 
-bus_1 = bus(name="Bus_1", wires=4) # Bus 1
+register_1 = register("A_register", 4, True, True) # A register
+register_2 = register("B_register", 4, True, True) # B register
 
-registers = [
-    register("A_register", 4, False, False, bus=bus_1),  # A register
-    register("B_register", 4, False, False, bus=bus_1),  # B register
-    register("Carry_register", 1, False, False),         # Carry register
-    register("Result_register", 4, False, False, bus=bus_1)  # Result register
-]
+register_3 = register("Carry_register", 1, True, True) # Carry register
+
+register_4 = register("Result_register", 4, True, True) # Result register
+
 
 def ALU(register_1, register_2, register_3, register_4):
     s = 0
@@ -17,24 +15,16 @@ def ALU(register_1, register_2, register_3, register_4):
 
     for i in range(4):
         print(f"\n\033[32mIteration {i+1}:\033[0m")
-
-        registers[0].enable_wire = True
-        for register in range(len(registers)):
-            a = int(registers[register].get()[3-i])
-        registers[0].enable_wire = False
-
-        registers[1].enable_wire = True
-        for register in registers:
-            b = int(register.get()[3-i])
-        registers[1].enable_wire = False
+        a = int(register_1.get()[3-i])
+        b = int(register_2.get()[3-i])
 
         print(f"\033[36m    a: {a} b: {b}\033[0m")
 
         s = XOR(XOR(a, b), c)
         c = OR(AND(a, b) , AND(c, XOR(a, b))) 
 
-        registers[3].modify(3-i, s)
-        registers[2].modify(0, c)
+        register_4.modify(3-i, s)
+        register_3.modify(0, c)
     
     return register_3.get(), register_4.get()
 
@@ -43,17 +33,17 @@ if __name__ == "__main__":
     test_counter = 0
     for i in range(16):
         for j in range(16):
-            registers[0].set(f"{i:04b}")
-            registers[1].set(f"{j:04b}")
-            registers[2].set("0")
-            registers[3].set("0000")
+            register_1.set(f"{i:04b}")
+            register_2.set(f"{j:04b}")
+            register_3.set("0")
+            register_4.set("0000")
 
             result = f"{i+j:04b}"
 
-            c, r = ALU(registers[0], registers[1], registers[2], registers[3])
+            c, r = ALU(register_1, register_2, register_3, register_4)
 
             print(f"\n\033[32mRight result: {result}\033[0m")
-            if len(result) > registers[0].bits or len(result) > register[1].bits and c == "1":
+            if len(result) > register_1.bits or len(result) > register_2.bits and c == "1":
                 color_code_1 = "\033[32m"
                 color_code_2 = "\033[0m"
             elif result == r and c == "0":
